CompareData: Stops: when at the same position, only consider equal when both gtfs & osm have the same type
Same type means either both are platforms, or both are stations

Otherwise, we may match a GTFS station with a OSM platform/stop, or a GTFS platform/stop with an OSM station
which would be wrong.

This was especially a problem when the GTFS station & GTFS stop/platform are
at the exact gps position.